+++ src/app/in/users/me/My-Account.tsx
@@ -1,7 +1,7 @@
 "use client";
 import React from "react";
 import Security from "@/components/Security2Fa/Security2Fa";
-import { useAuth } from "@/context/AuthContext-1";
+import { useAuth } from "@/context/AuthContext";
 import UpdateUserComponent from "@/components/Users/UpdateUser";
 
 const MyAccount = () => {
