+++ src/components/List-Invoice/ListInvoiceAsAdmin2.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { useAuth } from "@/context/AuthContext-1";
+import { useAuth } from "@/context/AuthContext";
 import React, { useEffect, useState } from "react";
 import { getAllInvoices, getUserById } from "@/helpers/auth.helper";
 import InvoiceDetail, {
