+++ src/components/dashboard/Avatar.tsx
@@ -2,7 +2,7 @@
 
 import React, { useState, useEffect } from "react";
 import Image from "next/image";
-import { useAuth } from "@/context/AuthContext-1";
+import { useAuth } from "@/context/AuthContext";
 
 const Avatar = () => {
   const { userData, loading } = useAuth();
