+++ src/components/deliverableWidget/DeliverableWidget.tsx
@@ -2,7 +2,7 @@
 "use client";
 import React, { useState, useEffect } from "react";
 
-import { useAuth } from "@/context/AuthContext-1";
+import { useAuth } from "@/context/AuthContext";
 import Cookies from "js-cookie";
 
 export interface Deliverable {
