+++ src/components/deliverablesActions/DeleteDeliverable.tsx
@@ -4,7 +4,7 @@ import React from "react";
 import Swal from "sweetalert2";
 import Cookies from "js-cookie";
 import { useRouter } from "next/navigation";
-import { useAuth } from "@/context/AuthContext-1";
+import { useAuth } from "@/context/AuthContext";
 
 const DeleteDeliverable = ({ id }: { id: any }) => {
   const token = Cookies.get("token");
