+++ src/components/deliverablesView/Searchbar.tsx
@@ -1,4 +1,4 @@
-import { useAuth } from "@/context/AuthContext-1";
+import { useAuth } from "@/context/AuthContext";
 import React, { useEffect, useState } from "react";
 import Cookies from "js-cookie";
