+++ src/components/deliverablesView/permissionpanel.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { useState, useEffect } from "react";
-import { AuthContextType, useAuth } from "@/context/AuthContext-1";
+import { AuthContextType, useAuth } from "@/context/AuthContext";
 import Swal from "sweetalert2";
 import Select from "react-select";
