+++ src/components/invoice/ButtonUploadInvoices.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import React from "react";
 import { PATHROUTES } from "@/helpers/pathRoutes";
-import { useAuth } from "@/context/AuthContext-1";
+import { useAuth } from "@/context/AuthContext";
 
 const ButtonUploadInvoice = ({ userId }: { userId: number }) => {
   const { userData } = useAuth();
