+++ src/components/sortDeliverables/SortDeliverable.tsx
@@ -2,7 +2,7 @@
 
 import React, { useState } from "react";
 import Cookies from "js-cookie";
-import { useAuth } from "@/context/AuthContext-1";
+import { useAuth } from "@/context/AuthContext";
 
 interface SortDeliverableProps {
   column: string;
